+++ packages/interface-internal/CHANGELOG.md
@@ -32,6 +32,36 @@
   * dependencies
     * @libp2p/peer-collections bumped from ^5.1.3 to ^5.1.4
 
+## [1.1.1](https://github.com/libp2p/js-libp2p/compare/interface-internal-v1.1.0...interface-internal-v1.1.1) (2024-04-24)
+
+
+### Documentation
+
+* fix broken links in docs site ([#2497](https://github.com/libp2p/js-libp2p/issues/2497)) ([fd1f834](https://github.com/libp2p/js-libp2p/commit/fd1f8343db030d74cd08bca6a0cffda93532765f)), closes [#2423](https://github.com/libp2p/js-libp2p/issues/2423)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.2.0 to ^1.3.0
+    * @libp2p/peer-collections bumped from ^5.1.10 to ^5.1.11
+
+## [1.1.0](https://github.com/libp2p/js-libp2p/compare/interface-internal-v1.0.11...interface-internal-v1.1.0) (2024-04-12)
+
+
+### Features
+
+* add isDialable method to libp2p ([#2479](https://github.com/libp2p/js-libp2p/issues/2479)) ([2c56203](https://github.com/libp2p/js-libp2p/commit/2c56203f9ccf4b6ed30541a871b9bd8c5a21526e))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.1.6 to ^1.2.0
+    * @libp2p/peer-collections bumped from ^5.1.9 to ^5.1.10
+
 ## [1.0.11](https://github.com/libp2p/js-libp2p/compare/interface-internal-v1.0.10...interface-internal-v1.0.11) (2024-04-05)