+++ packages/logger/CHANGELOG.md
@@ -21,6 +21,33 @@
   * devDependencies
     * @libp2p/peer-id bumped from ^4.0.2 to ^4.0.3
 
+## [4.0.11](https://github.com/libp2p/js-libp2p/compare/logger-v4.0.10...logger-v4.0.11) (2024-04-24)
+
+
+### Documentation
+
+* fix broken links in docs site ([#2497](https://github.com/libp2p/js-libp2p/issues/2497)) ([fd1f834](https://github.com/libp2p/js-libp2p/commit/fd1f8343db030d74cd08bca6a0cffda93532765f)), closes [#2423](https://github.com/libp2p/js-libp2p/issues/2423)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.2.0 to ^1.3.0
+  * devDependencies
+    * @libp2p/peer-id bumped from ^4.0.10 to ^4.1.0
+
+## [4.0.10](https://github.com/libp2p/js-libp2p/compare/logger-v4.0.9...logger-v4.0.10) (2024-04-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.1.6 to ^1.2.0
+  * devDependencies
+    * @libp2p/peer-id bumped from ^4.0.9 to ^4.0.10
+
 ## [4.0.9](https://github.com/libp2p/js-libp2p/compare/logger-v4.0.8...logger-v4.0.9) (2024-04-05)