+++ packages/transport-webtransport/CHANGELOG.md
@@ -96,6 +96,66 @@
   * devDependencies
     * libp2p bumped from ^1.2.2 to ^1.2.3
 
+## [4.0.28](https://github.com/libp2p/js-libp2p/compare/webtransport-v4.0.27...webtransport-v4.0.28) (2024-04-24)
+
+
+### Documentation
+
+* fix broken links in docs site ([#2497](https://github.com/libp2p/js-libp2p/issues/2497)) ([fd1f834](https://github.com/libp2p/js-libp2p/commit/fd1f8343db030d74cd08bca6a0cffda93532765f)), closes [#2423](https://github.com/libp2p/js-libp2p/issues/2423)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.2.0 to ^1.3.0
+    * @libp2p/peer-id bumped from ^4.0.10 to ^4.1.0
+  * devDependencies
+    * @libp2p/logger bumped from ^4.0.10 to ^4.0.11
+    * @libp2p/peer-id-factory bumped from ^4.0.10 to ^4.1.0
+    * libp2p bumped from ^1.4.2 to ^1.4.3
+
+## [4.0.27](https://github.com/libp2p/js-libp2p/compare/webtransport-v4.0.26...webtransport-v4.0.27) (2024-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * libp2p bumped from ^1.4.1 to ^1.4.2
+
+## [4.0.26](https://github.com/libp2p/js-libp2p/compare/webtransport-v4.0.25...webtransport-v4.0.26) (2024-04-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * libp2p bumped from ^1.4.0 to ^1.4.1
+
+## [4.0.25](https://github.com/libp2p/js-libp2p/compare/webtransport-v4.0.24...webtransport-v4.0.25) (2024-04-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.1.6 to ^1.2.0
+    * @libp2p/peer-id bumped from ^4.0.9 to ^4.0.10
+  * devDependencies
+    * @libp2p/logger bumped from ^4.0.9 to ^4.0.10
+    * @libp2p/peer-id-factory bumped from ^4.0.9 to ^4.0.10
+    * libp2p bumped from ^1.3.3 to ^1.4.0
+
+## [4.0.24](https://github.com/libp2p/js-libp2p/compare/webtransport-v4.0.23...webtransport-v4.0.24) (2024-04-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * libp2p bumped from ^1.3.2 to ^1.3.3
+
 ## [4.0.23](https://github.com/libp2p/js-libp2p/compare/webtransport-v4.0.22...webtransport-v4.0.23) (2024-04-05)